Add a write test for the health-check
This allows us to read/write a specified/random value.

We really need to make sure that the person does not overwrite
anything in the data store via HealthCheck so we need that extra
check. I know we could just generate another random val, but the
current solution seems like it does the job in preventing overwrites.